backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0
asyncpg==0.29.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
httpx==0.27.0
